Removed exporter folder and refactored code to build without it
