+++ charts/brigade-noisy-neighbor/values.yaml
@@ -15,7 +15,7 @@ brigade:
   ## (http:// or https://)
   apiAddress: https://brigade-apiserver.brigade.svc.cluster.local
   ## API token belonging to a Brigade 2 service account
-  apiToken: 5e0a44265e6b4867a7c3535ef2f01b54E8YF8RtFk7psuAAOaOMKqbWnlJYBThFSz0n9ahLPqeIGeJzoRTmqLGpApQojAZz0Tvy8tpb1aPcwdAuUdbP41YyJxfioHx4fW9qEaRpdrloWPI5DboOtN5hXj3Ja1e3UkQOGlpVbPFUgopGgtXFYMjZW9ROw89r9JWEtMauwxmkacbVQrRqHbQwBWrCsra79zdBZEfMBriWdAcO0MRKe8Kd8k8LA9AON
+  apiToken: <placeholder>
   ## Whether to ignore cert warning from the API server
   apiIgnoreCertWarnings: true
