+++ internal/signals/signals_test.go
